sort of works; minimongo is super sluggish
