+++ .meteor/versions
@@ -22,9 +22,9 @@ clinical:extended-api@2.5.0
 clinical:glass-ui@2.5.0
 clinical:hl7-resource-bundle@1.5.5
 clinical:hl7-resource-datatypes@4.0.5
-clinical:hl7-resource-encounter@2.0.3
+clinical:hl7-resource-encounter@2.0.4
 clinical:hl7-resource-endpoint@3.3.2
-clinical:hl7-resource-patient@5.0.5
+clinical:hl7-resource-patient@5.0.6
 clinical:user-model@1.7.0
 coffeescript@1.0.17
 ddp@1.4.0
@@ -81,6 +81,7 @@ npm-bcrypt@0.9.3
 npm-mongo@3.1.2
 observe-sequence@1.0.16
 ordered-dict@1.1.0
+patientinsight:cardiac-scorecard@0.2.2
 promise@0.11.2
 raix:eventemitter@0.1.3
 random@1.1.0
@@ -105,7 +106,6 @@ standard-minifier-css@1.5.3
 standard-minifier-js@2.4.1
 static-html@1.2.2
 symptomatic:care-cards@0.4.0
-symptomatic:nivo-analytics@0.2.0
 templating@1.2.15
 templating-compiler@1.2.15
 templating-runtime@1.2.15
