+++ package.json
@@ -42,8 +42,9 @@
     "bunyan": "1.8.12",
     "classnames": "2.2.6",
     "clsx": "1.0.4",
+    "fhir-kit-client": "1.4.0",
     "lodash": "^4.17.11",
-    "material-fhir-ui": "0.7.6",
+    "material-fhir-ui": "0.7.10",
     "material-ui": "^0.20.2",
     "meteor-node-stubs": "^0.4.1",
     "minimongo": "6.3.1",
